Tweaked tags and end loop sequence
Changed all tags and variables to have "gp_" at the beginning to differ them from the wandering trader functions.

I also changed the scoreboard reset for the variables to ensure it doesn't end the trade adding early.

I removed the math folder as it is obsolete since the random command has been added